requirements.txt
discord.py[voice]>=2.4.0
yt-dlp>=2025.1.1
spotipy==2.24.0
python-dotenv==1.0.1
PyNaCl==1.5.0
aiohttp>=3.9.5
static-ffmpeg>=2.5
